Remove check_required_components() macro
Should we add support to find_package() to be found by COMPONENTS
option? The packages that is needed is already has to be configured when
building, later this file makes calls to find_depedency(). Kind of sound
like chicken egg problem; to be able to build, say, OpenGL backend, the
OpenGL needed to be found (called with REQUIRED option on
find_package()). Does it has any meaning to passing COMPONENTS OpenGL 3?

Also the current CMake support makes no distinction between platform and
backend, the FreeGLUT, GLFW3, SDL(2|3) are platforms, OpenGL(2|3),
Vulkan, Metal are backends. I choosed to not to make distinction because
upstream ImGui doesn't make so. All the related code on backends/
folder. Anyway... :)