Fix link error on example SDL2+WGPU on Windows
wgpu-native backends is still not a smooth experience on Windows. It
needs a find_library() call which finds shared libraries by default but
that file needs to copy on the built examples' executable folder. When
the CMAKE_FIND_LIBRARY_SUFFIXES tweak to find static libraries, some
extra symbol needed to be link.